Move ProjectSchema calculation into Gradle plugins classloader
