+++ provider-plugins/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.PluginServiceRegistry
@@ -1,0 +1,1 @@
+org.gradle.kotlin.dsl.provider.plugins.KotlinDslProviderPluginsServiceRegistry
